Create game_ops.rb file
